backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.10.2
pydantic-settings==2.6.1

# OpenAI for embeddings and LLM
openai==1.57.0

# Qdrant client for vector database
qdrant-client==1.12.1

# PostgreSQL async driver
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0

# Rate limiting
slowapi==0.1.9
limits==3.13.0

# HTTP client
httpx==0.28.0
aiofiles==24.1.0

# Logging and monitoring
structlog==24.4.0
python-json-logger==2.0.7

# Development dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1

# Environment variables
python-dotenv==1.0.1